# refactor: Use typeMismatchError rather than raw string error